Unit test mbedtls_timing_hardclock
Added a test for mbedtls_timing_hardclock to check that its rate is
approximately constant.

This test fails under moderate to high load, and could also
legitimately fail on a system where the function uses a cycle counter
and the processor's clock rate is variable. So I don't think this is a
good idea.