Timing: align return value when resetting a timer
When mbedtls_timing_get_timer is called with a nonzero reset argument,
the Windows implementation of returned the previous elapsed time, but
the Unix/POSIX implementation returned 0. The documentation did not
make it clear what the return value should be in this case (0 is the
elapsed time after resetting the timer). Align with the Windows
implementation as this is arguably somewhat useful.