mbedtls_timing_get_timer: return 0 on reset
Change mbedtls_timing_get_timer to return 0 on reset on both Windows
and Unix. This functionally reverts the previous commit
"Timing: align return value when resetting a timer": the behavior is
now aligned with the original Windows behavior rathen than with the
original Unix behavior.

The reason for this revert is that mbedtls_timing_get_timer(timer, 1)
is the way to initialize a timer. It must not access the timer's
fields since they may not initialized yet. This commit fixes the
Windows behavior as well: it was accessing the potentially
uninitialized values. The previous code could have crashed with a
toolchain that enabled signed integer overflow detection, but in most
setup it just returned undefined values.