SSL reproducible test mode
New option fake_entropy for ssl_server2 and ssl_client2. With this
option, the programs act deterministically based on the current time.
This is intended for developer testing.

To get deterministic behavior, pass fake_entropy=foo and either unset
MBEDTLS_HAVE_TIME in the configuration or use a tool such as faketime
to set a constant time, e.g.

    faketime -f "2015-12-22 00:35:08" programs/ssl/ssl_server2 server_addr=127.0.0.1 fake_entropy=foo