Support larger work factors
Support arbitrarily large work factors up to the size of unsigned int.
The previous implementation was limited by the size of bignums.

Since the exponent in mbedtls_mpi_exp_mod has to be represented as a
bignum, do the exponentiation in several steps, with a good compromise
between performance and memory usage.