udp_proxy: add support for packet modifications (xor)
Add support for applying modifications (patches) to specific packets.
It is possible to apply multiple modifications successively. A
modification applies to a specific packet (numbered from 1) in a
specific direction (C->S or S->C).

Currently the only supported type of modification is to xor a string
with the packet. This allows corrupting a packet providing that the
bits to corrupt are at a known location in the conversation.