In deterministic tests, don't fall back to randomness
Change rnd_buffer_rand to not fall back to rnd_std_rand if rng_state
is null or if the preset buffer of fake random data has run out. If a
test is deterministic, it needs the full buffer length. If a test is
not deterministic, it should call rnd_std_rand or one of the DRBG
modules.

This avoid accidentally writing test data that actually depends on the
platform's rand implementation.