Test precise error codes from mbedtls_ecdsa_read_signature
When testing crafted invalid signatures, check the precise error code,
not just != 0.

The case with MBEDTLS_ERR_ECP_SIG_LEN_MISMATCH implements the
documented behavior, which is not the actual behavior of the code. The
documented behavior is to return this error if the buffer contains a
valid signature plus trailing garbage. The actual behavior is to
return this error if the buffer contains an ASN.1 SEQUENCE (plus
nothing else) which inside contains r, s and some trailing garbage;
this case is currently untested, and doesn't sound useful.