Change the reporting format to be compiler-like
Instead of reporting issues separately with an ad hoc format, report
them in a grep-like format, grouped by file and then by line. The
grouping makes it easier to fix multiple issues in the same or nearby
locations. The grep-like format allows using IDE automation to
navigate through the output as if it was compiler output.

Internally, this commit changes the way issues are collected: instead
of collecting them and reporting them independently in each tracker,
collect them in a global dictionary with an attached description, and
report them together.