Remove timing self-test
Self-tests are useful in cryptography modules because many
certification schemes require them. For an operating system interface
abstraction layer, self-tests aren't useful.

Furthermore the self-tests in the timing module are unreliable: they
depend on how fast the operating system executes the module.
Specifically, they fail if the process remains unscheduled for too
long, even though the code is working as intended.

So remove the self-tests from the timing module altogether.