all.sh: test ssl3 in the full-config build
Since SSL3 is back in the full-config build, there's no need to test
it in a separate component. (It was a separate component even before
SSL3 was removed from the full-config build, but that was not necessary.)

Test SSLv3 interoperability with GnuTLS as well. For whatever reason,
there was only interoperability with OpenSSL.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>