Add a proper mechanism to skip demos with unmet dependencies
Declare the dependencies of a demo script in the script itself. If the
script uses multiple programs or uses programs in a specific way, it
may have more dependencies than the program(s) itself.

Call config.py as a subprocess so that this script can be easily
backported to branches that only have config.pl.

Print a short message if a demo script is skipped due to an unmet dependency.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>