Exclude DEPRECATED_WARNING in config full
In the full config, don't set MBEDTLS_DEPRECATED_WARNING. This is debatable:
the full config does not enable deprecated features in this branch, so
MBEDTLS_DEPRECATED_WARNING is compatible with the other features.
Exclude it to keep LTS branches closer to development.

In any case, baremetal and full should have the same settings regarding
deprecated features, so don't do anything about DEPRECATED_xxx in baremetal.

In all.sh, test MBEDTLS_DEPRECATED_WARNING through the new component
component_test_full_deprecated_warning. This is similar, but not identical to
the component_test_full_no_deprecated_warning in development. The difference
is due to development having deprecated features in the "full" configuration.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>