Include <limits.h> where needed
Include <limits.h> in test suites and sample or test programs that use
it but didn't include it explcitly.

This will be necessary for a subsequent commit that causes config.h to
no longer include <limits.h>.

Remove the inclusion from test_suite_timing which stopped requiring it
a while ago.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>