Forbid 16-bit platforms
Refuse to compile the library if size_t is less than 32 bits, because
some of the code is known to be vulnerable to integer overflows even
though it would compile cleanly (TLS code adding 16-bit or 24-bit
values).

Refuse to compile the library if size_t is less than int (and in
particular if size_t is less than 32 bits).

Also refuse to compile the library if int isn't at least 32 bits.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>