Remove the obsolescent configs/config-psa-crypto.h
The sample configuration file config-psa-crypto.h was introduced
during the early development of PSA crypto. Now that
MBEDTLS_PSA_CRYPTO_C is in the default configuration, it is no longer
useful. This file has not been kept synchronized with config.h lately,
so its content is no longer up to date.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>