Gather simple statistics about dynamic memory allocations
When MBEDTLS_TEST_MEMORY_WRAPPERS is enabled, gather some simple
statistics about memory allocations in each test case and print them
out.

So far the information is only reported in the host test runner, not
in the on-target test runner.

Also include a simple memory leak detector: consider it an error in
the test suite if the number of calls to mbedtls_free() is less than
the number of calls to mbedtls_calloc().

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>