New function reset_test_failure()
Put the code to reset the test failure information between tests in a
single function.

Call this function just before executing a test, instead of having
separate code for it in the host runner and in the on-target runner.
This incidentally fixes a bug that the test step information wasn't
reset between tests with the on-target runner.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>