Call calloc/free wrappers in tests to match free/calloc in the library
The calloc() and free() wrappers defined in test
code (mbedtls_test_calloc_wrapper and mbedtls_test_calloc_free) are meant to
be used only for allocations made inside the library, and not for
allocations of test data. This is because they are meant to measure or
instrument what the library is doing, not what the tests are doing.

Occasionally a direct call to mbedtls_free() in test code frees an object
that was allocated inside a library function, or a direct call to
mbedtls_calloc() in test code allocates an object that will be freed inside
a library function. In these cases, do call the wrapper function from the
test. This way the allocation counts in the wrappers will balance.

This fixes the leaks, and negative numbers of leaks, that the tests were
reporting with the wrappers enabled.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>