Merge SSL buffer shrinking code with handle_buffer_resizing
The buffer shrinking code in mbedtls_ssl_handhake_free() was identical
to the enlargement code that a previous commit moved from
ssl_handshake_init() to handle_buffer_resizing(), except for different
values of {in,out}_buf_new_len and for different conditions to trigger
the resizing. Add a parameter to handle_buffer_resizing() to select
between the two sets of buffer resizing conditions.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>