all.sh; only run pylint if available
Only run pylint if it's available and recent enough. If pylint is
missing or too old, don't list check_python_files as a supported
component.

This way, given the state of our CI as of writing, check_python_files
will run on the Jenkins Ubuntu 18.04 system and on Travis but not on
the Jenkins Ubuntu 16.04 system which is using the Ubuntu package
which no longer works.

Pylint3 1.5.2 from Ubuntu 16.04 no longer works because it doesn't
support the init-hook option, and so it doesn't add the scripts
directory to the module search path, leading to the error
`Unable to import 'mbedtls_dev'`.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>